- Split the Configure Diff Tool section into its own page in the Settings screen.
