+++ VS.DiffAllFiles/Settings/DiffAllFilesSettingsPageControl.xaml
@@ -6,7 +6,7 @@
 						 xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit"
 						 xmlns:QC="clr-namespace:QuickConverter;assembly=QuickConverter"
 						 mc:Ignorable="d" 
-						 d:DesignHeight="350" d:DesignWidth="400" LostKeyboardFocus="UserControl_LostKeyboardFocus">
+						 d:DesignHeight="310" d:DesignWidth="400" LostKeyboardFocus="UserControl_LostKeyboardFocus">
 	<UserControl.Resources>
 		<Style TargetType="{x:Type Button}">
 			<Setter Property="Padding" Value="5" />
@@ -40,7 +40,7 @@
 					<TextBox Text="{Binding Path=FileExtensionsToIgnore}" HorizontalAlignment="Stretch" ToolTip="File types to not compare." />
 				</StackPanel>
 			</GroupBox>
-			
+
 			<GroupBox Header="How To Compare Files" Margin="0,10,0,0">
 				<StackPanel Orientation="Vertical">
 					<RadioButton GroupName="HowToCompareFiles" Content="Always compare each file independently" IsChecked="{QC:Binding '$P == CompareModes.IndividualFiles', P={Binding Path=CompareModesAvailable, Mode=OneTime}}" Checked="radioCompareMode_Independent_Checked" ToolTip="If selected, each file will be compared independently in its own window." />
@@ -55,7 +55,7 @@
 							</Label.ToolTip>
 						</Label>
 					</StackPanel>
-					
+
 					<RadioButton GroupName="HowToCompareFiles" Content="Always combine and compare all files in a single window" IsChecked="{QC:Binding '$P == CompareModes.CombinedIntoSingleFile', P={Binding Path=CompareModesAvailable, Mode=OneTime}}" Checked="radioCompareMode_Combined_Checked" ToolTip="If selected, all files will be combined and compared in a single window." Margin="0,10,0,0" />
 					<StackPanel Orientation="Horizontal" Margin="30,0,0,0">
 						<CheckBox Content="Use same file headers in source and target files" IsChecked="{Binding Path=UseSameHeadersForCombinedFiles}">
@@ -68,18 +68,11 @@
 							</CheckBox.ToolTip>
 						</CheckBox>
 					</StackPanel>
-					
+
 					<RadioButton GroupName="HowToCompareFiles" Content="Allow me to decide by showing a checkbox on the UI" IsChecked="{QC:Binding '$P == CompareModes.AllowUserToChoose', P={Binding Path=CompareModesAvailable, Mode=OneTime}}" Checked="radioCompareMode_LetUserDecide_Checked" ToolTip="If selected, a checkout will be shown on the UI allowing you to specify which compare mode to use." Margin="0,10,0,0" />
 				</StackPanel>
 			</GroupBox>
-			
-			<StackPanel Orientation="Horizontal" VerticalAlignment="Center" Margin="0,10,0,0">
-				<Button Content="Configure Diff Tool..." HorizontalAlignment="Left" Click="btnConfigureDiffTool_Click" ToolTip="Select the tool to use for comparing different types of files." />
-				<TextBlock VerticalAlignment="Center" Margin="5,0,0,0">
-						<Hyperlink NavigateUri="https://diffallfilesvisualstudioextension.codeplex.com/wikipage?title=Common%20Diff%20Tool%20Configurations%20For%20Visual%20Studio" RequestNavigate="Hyperlink_RequestNavigate">View common diff tool configurations</Hyperlink>
-				</TextBlock>
-			</StackPanel>
-			
+
 			<Button Content="Restore Default Settings" HorizontalAlignment="Right" Margin="0,10,0,0" Click="btnRestoreDefaultSettings_Click" />
 		</StackPanel>
 	</Grid>
