+++ VS.DiffAllFiles/VS.DiffAllFilesPackage.cs
@@ -21,8 +21,9 @@ namespace VS_DiffAllFiles
     [PackageRegistration(UseManagedResourcesOnly = true)]
     // This attribute is used to register the information needed to show this package in the Help/About dialog of Visual Studio.
     [InstalledProductRegistration("#110", "#112", "1.0", IconResourceID = 400)]
-	// Add our DiffAllFilesSettings class to the Tools -> Options menu.
+	// Add our Diff All Files section to the Tools -> Options menu.
 	[ProvideOptionPage(typeof(DiffAllFilesSettings), "Diff All Files", "General", 0, 0, true)]
+	[ProvideOptionPage(typeof(ConfigureDiffToolSettingsPage), "Diff All Files", "Diff Tool", 0, 0, true)]
 	// Auto Load our assembly even when no solution is open (by using the Microsoft.VisualStudio.VSConstants.UICONTEXT_NoSolution guid).
 	[ProvideAutoLoad("ADFC4E64-0397-11D1-9F4E-00A0C911004F")]
 	// Include other assemblies in the same directory as this package that this package depends on (e.g. WPF Extended Toolkit, QuickConverter, etc.).
